login virker med token

+++ NGK3Assignment/Controllers/AccountController.cs
@@ -54,7 +54,7 @@ namespace NGK3Assignment.Controllers
 
         // GET: api/Account/5
         [HttpGet("{id}", Name = "Get")]
-        public async Task<ActionResult<UserDto>> Get(int id)
+        public async Task<ActionResult<UserDto>> Get(long id)
         {
             var user = await _context.Users.FindAsync(id);
             if (user == null)
